Update
Delete console logs

Create separate file for new functions in main
